[MER-5396] Extract Canvas grade passback test helpers
